pyproject.toml
[tool.poetry]
name = "ml_engineering_lab3"
description = "Training pipeline for deep learning model"
authors = ["amex.ua@hotmail.com"]
version = "0.01"
package-mode = false

[tool.poetry.dependencies]
python = "~3.11"
dvc = "^3.59"
torch = { version = "^2.6.0+cu118", source = "pytorch" }
torchvision = { version = "^0.21.0+cu118", source = "pytorch" }
scikit-learn = "^1.6"
tqdm = "^4.67"
matplotlib = "^3.10"
numpy = "^2.2"
pyyaml = "^6.0"
scipy = "^1.15"
pandas = "^2.2"
requests = "^2.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.15"
ruff = "^0.11"
black = "^25.0"
isort = "^6.0"
ipykernel = "^6.29"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch"
priority = "supplemental"
url = "https://download.pytorch.org/whl/cu118"
